feat: enforce shape on Services + rename
