+++ src/core/database/BaseService.ts
@@ -2,7 +2,7 @@ import mongoose from 'mongoose';
 import { IDocument } from '../library/types/Miscellaneous';
 import { Database } from './Database';
 import { ServiceMetadataKey } from './decorators/DatabaseService';
-import { Services } from './types/ServiceMap';
+import { Services } from './types/Services';
 import { ModelMetadataKey } from './decorators/DatabaseModel';
 
 export abstract class BaseService<D extends IDocument = IDocument> {
