+++ src/core/database/decorators/DatabaseService.ts
@@ -1,6 +1,6 @@
 import { ConstructorFunction } from '../../library/types/Miscellaneous';
 import { BaseService } from '../BaseService';
-import { ServiceMapKeys } from '../types/ServiceMap';
+import { ServiceMapKeys } from '../types/Services';
 
 export const ServiceMetadataKey = Symbol('db:serviceKey');
