+++ src/core/database/services/Users.ts
@@ -17,7 +17,7 @@ export class Users<D extends IUser = IUser> extends BaseService<D> {
 }
 
 /* Declare Users to augment the ServiceMap */
-declare module '../types/ServiceMap' {
+declare module '../types/Services' {
   interface Services {
     users: Users;
   }
