+++ src/core/database/types/Services.ts
@@ -5,7 +5,7 @@ export type ServiceMapConstraint = Record<string, BaseService<any>>;
 
 // This interface can be augmented via declaration merging
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Services {}
+export interface Services extends ServiceMapConstraint {}
 
 // Helper type to extract service keys
 export type ServiceMapKeys = keyof Services;
